Restructure repository.
The repository root is the Racket package.

The Python code is in `py/pydrnlp/`.
A Conda environment is expected at `py/condaenv/` based on
the specification at `py/requirements.yaml`.
In the future, it would be good to have Racket set
this up automatically.
I am trying Conda for now because it provides `spaCy`
pre-built, but I haven't ruled out using venv.

The directory `py/compare-env-vars/` contains Racket
scripts and data that were used interactively
to see what environment variables are set by activating
a Conda or venv virtual environment.

The `old/` directory contains various things
that haven't been updated based on the restructuring done in this
commit. It should be cleaned up further.
Some of those things may not logically belong in this repository.