edit history to remove giant sqlite db
Creating the GitHub mirror caused a warning
that `trends/tokenizer-cache-db.sqlite` had been checked in
and was 57.02 MB.

It wasn't supposed to have been checked in at all.

Prior to this commit, I edited the repository's history to remove it.
The history is saved at <https://bitbucket.org/philipmcgrath/pydrnlp-history>,
just in case.

This commit adds the BFG report details, temporarily, to make
them available for future reference in the unlikely event they're
ever needed.